refactor: Unify Dual HVAC Modules (ARCH-003)
- Move src/hvac/ to src/sim/hvac/zones/
- Rename control.rs to modes.rs
- Update all import paths
- Remove top-level hvac module from lib.rs

Closes #623
