+++ src/lib.rs
@@ -67,7 +67,6 @@ pub mod ai;
 pub mod analysis;
 pub mod api;
 pub mod cli;
-pub mod hvac;
 pub mod interop;
 pub mod napi;
 pub mod performance;
