+++ src/sim/hvac/zones/schedule.rs
